Erro 404 da entidade Cliente resolvido.

+++ Backend/cliente/src/main/java/com/cliente/demo/model/Cliente.java
@@ -17,7 +17,7 @@ public class Cliente {
     private Integer id;
     
     private String nome;
-    private Integer telefone;
+    private Long telefone;
 
     public Cliente() {
     }
